# docs/swagger.yaml
basePath: /api/v1
definitions:
  admin_dto.AdminCreateStudentAccountInput:
    properties:
      code:
        type: string
      email:
        type: string
      name:
        type: string
      phone_number:
        type: string
      sub_major_id:
        type: integer
    required:
    - code
    - email
    - name
    - phone_number
    - sub_major_id
    type: object
  admin_dto.AdminCreateTeacherAccountInput:
    properties:
      email:
        type: string
      name:
        type: string
      phone_number:
        type: string
      sub_major_id:
        type: integer
    required:
    - email
    - name
    - phone_number
    - sub_major_id
    type: object
  auth_dto.ForgotPasswordInput:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  auth_dto.LoginInput:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  auth_dto.LoginOutput:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
  auth_dto.LoginSwaggerOutput:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/auth_dto.LoginOutput'
      message:
        type: boolean
    type: object
  auth_dto.ResetPasswordInput:
    properties:
      password:
        type: string
      token:
        type: string
    required:
    - password
    - token
    type: object
  dto.MetaPagination:
    properties:
      current_page:
        type: integer
      total:
        type: integer
    type: object
  major_dto.GetListMajorOutput:
    properties:
      items:
        items:
          $ref: '#/definitions/major_dto.MajorOutput'
        type: array
      meta:
        $ref: '#/definitions/dto.MetaPagination'
    type: object
  major_dto.GetMajorSwaggerOutput:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/major_dto.MajorOutput'
      message:
        type: boolean
    type: object
  major_dto.MajorOutput:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  response.ResponseDataSuccess:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: boolean
    type: object
  response.ResponseErr:
    properties:
      code:
        type: integer
      error: {}
      message:
        type: boolean
    type: object
  sub_major_dto.GetListSubMajorOutput:
    properties:
      items:
        items:
          $ref: '#/definitions/sub_major_dto.SubMajorOutput'
        type: array
      meta:
        $ref: '#/definitions/dto.MetaPagination'
    type: object
  sub_major_dto.GetSubMajorSwaggerOutput:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/sub_major_dto.SubMajorOutput'
      message:
        type: boolean
    type: object
  sub_major_dto.SubMajorOutput:
    properties:
      id:
        type: integer
      major_id:
        type: integer
      name:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: minhbeardev@gmail.com
  description: Capstone Project FPT API
  title: Capstone Project FPT API
  version: "1.0"
paths:
  /admin/students/create-account:
    post:
      consumes:
      - application/json
      description: Admin Create Student Account
      parameters:
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/admin_dto.AdminCreateStudentAccountInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseDataSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ResponseErr'
      security:
      - ApiKeyAuth: []
      summary: CreateStudentAccount
      tags:
      - Admin
  /admin/students/import-data:
    post:
      consumes:
      - multipart/form-data
      description: Admin upload Excel file to create student accounts
      parameters:
      - description: Excel file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseDataSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ResponseErr'
      security:
      - ApiKeyAuth: []
      summary: UploadFileStudentData
      tags:
      - Admin
  /admin/teachers/create-account:
    post:
      consumes:
      - application/json
      description: Admin Create Teacher Account
      parameters:
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/admin_dto.AdminCreateTeacherAccountInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseDataSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ResponseErr'
      security:
      - ApiKeyAuth: []
      summary: CreateTeacherAccount
      tags:
      - Admin
  /admin/teachers/import-data:
    post:
      consumes:
      - multipart/form-data
      description: Admin upload Excel file to create teacher accounts
      parameters:
      - description: Excel file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseDataSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ResponseErr'
      security:
      - ApiKeyAuth: []
      summary: UploadFileTeacherData
      tags:
      - Admin
  /forgot-password:
    post:
      consumes:
      - application/json
      description: Forgot Password
      parameters:
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/auth_dto.ForgotPasswordInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseDataSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ResponseErr'
      summary: ForgotPassword
      tags:
      - Auth
...
  /login:
    post:
      consumes:
      - application/json
      description: Login
      parameters:
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/auth_dto.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth_dto.LoginSwaggerOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ResponseErr'
      summary: Login
      tags:
      - Auth
  /login/google:
    post:
      description: Login With Google
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseDataSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ResponseErr'
      summary: Login With Google
      tags:
      - Auth
  /majors:
    get:
      consumes:
      - application/json
      description: Get list major
      parameters:
      - description: Limit
        in: query
        name: limit
        required: true
        type: integer
      - description: Page
        in: query
        name: page
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/major_dto.GetListMajorOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ResponseErr'
      summary: GetListMajor
      tags:
      - Public
  /majors/{id}:
    get:
      description: Get Major
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/major_dto.GetMajorSwaggerOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ResponseErr'
      security:
      - ApiKeyAuth: []
      summary: GetMajor
      tags:
      - Public
  /reset-password:
    post:
      consumes:
      - application/json
      description: Reset Password
      parameters:
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/auth_dto.ResetPasswordInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseDataSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ResponseErr'
      summary: ResetPassword
      tags:
      - Auth
  /sub-majors:
    get:
      consumes:
      - application/json
      description: Get list sub major
      parameters:
      - description: Limit
        in: query
        name: limit
        required: true
        type: integer
      - description: Page
        in: query
        name: page
        required: true
        type: integer
      - description: Major ID
        in: query
        name: major_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sub_major_dto.GetListSubMajorOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ResponseErr'
      summary: GetListSubMajor
      tags:
      - Public
  /sub-majors/{id}:
    get:
      description: Get Sub Major
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sub_major_dto.GetSubMajorSwaggerOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ResponseErr'
      security:
      - ApiKeyAuth: []
      summary: GetSubMajor
      tags:
      - Public
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
  BasicAuth:
    type: basic
swagger: "2.0"